Empty Clip Fix
Addresses the issue where empty clips couldn't be crafted. They DO NOT WORK currently. I believe the state of the items messed up because of how /empty is defined. They appear to be full, but do not work in any capacity. Pushing so they can atleast be crafted. Will come back around to this.